// src/types/vendor.ts
export type Vendor = {
  _id?: string
  name?: string
  phone_no?: string
  address?: string
  email?: string
  note?: string
  photo?: any
  updatedAt?: string
  createdAt?: string
  deletedAt?: string
  balance_amount?: number
  last_paid_amount?: number
  last_paid?: string
  search?: string
  status?: string
}